Move the low-level wrapper for cm shell into a dedicated file
With all the internal static variables, only exposing the minimal set of free functions:

- Launch()     (was LaunchBackgroundPlasticShell())
- Terminate()
- RunCommand() (was RunCommandInternal())

Now Utils only contains functions on a higher level, calling RunCommandRaw() and parsing the result
